refactor: put helper function in index.ts--add const for routes
