+++ src/app/router/routes.ts
@@ -2,5 +2,7 @@ const HOME_ROUTE = "/";
 const POOLS_ROUTE = "pools";
 const ADD_LIQUIDITY = "add-liquidity";
 const SWAP_ROUTE = "swap";
+const POOLS_ADD_LIQUIDITY = "/pools/add-liquidity";
+const POOLS_PAGE = "/pools";
 
-export { HOME_ROUTE, POOLS_ROUTE, SWAP_ROUTE, ADD_LIQUIDITY };
+export { HOME_ROUTE, POOLS_ROUTE, SWAP_ROUTE, ADD_LIQUIDITY, POOLS_ADD_LIQUIDITY, POOLS_PAGE };
