--- src/helper/formatInputTokenValue.ts
@@ -1,3 +1,0 @@
-export const formatInputTokenValue = (base: number, exponent: string) => {
-  return base * Math.pow(10, parseFloat(exponent));
-};
