+++ src/helper/index.ts
@@ -6,3 +6,12 @@ export const reduceAddress = (address: string | undefined, lengthLeft: number, l
   }
   return "Not connected";
 };
+
+export const calculateSlippage = (base: string, percent: number | undefined = 0) => {
+  const result = (parseFloat(base) * percent) / 100;
+  return Math.ceil(result);
+};
+
+export const formatInputTokenValue = (base: number, exponent: string) => {
+  return base * Math.pow(10, parseFloat(exponent));
+};
